adding getHash method to every pojo

+++ src/main/java/io/kebblar/petstore/api/model/domain/Rol.java
@@ -155,7 +155,7 @@ public class Rol {
      *
      * @return entero representando el hash
      */
-    public long getHash() {
+    public int getHash() {
         return hashCode();
     }
 }
